+++ crux-llvm/CHANGELOG.md
@@ -1,6 +1,7 @@
 # next -- TBA
 
 * Add support for the Bitwuzla SMT solver in the test suite.
+* Add `--debug` option for starting the Crucible debugger.
 
 # 0.9 -- 2024-08-30